+++ components/Button.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Link from 'next/link'
 
 export default function Button(props) {
-  const { hasToggle, name, nextPage, size, type, onSubmit } = props;
+  const { hasToggle, name, nextPage, size, type, onClick } = props;
 
   const buttonColor = name == 'back' ? 'danger' : 'primary'
   const buttonType = type == 'submit' ? 'submit' : 'button'
@@ -14,7 +14,7 @@ export default function Button(props) {
           title={name}
           type={buttonType}
           disabled={hasToggle}
-          onClick={onSubmit}
+          onClick={onClick}
         >
           {name}
         </button>