+++ components/contacts/utils/messageInitialValues.js
@@ -1,0 +1,5 @@
+export const messageInitialValues = {
+    completeName: "",
+    email: "",
+    description: ""
+  }