+++ utils/validation.js
@@ -1,5 +1,6 @@
-//Philippine Mobile Phone Number Regex pattern: /((\+639)|09)\d{9}/;
 import * as Yup from "yup"; //import yup
+
+//Philippine Mobile Phone Number Regex pattern: /((\+639)|09)\d{9}/;
 
 export const validationSchema = Yup.object({
   firstName: Yup.string()
@@ -11,17 +12,20 @@ export const validationSchema = Yup.object({
   lastName: Yup.string()
     .max(50, "Last name is too long")
     .required("This field is required"),
-  gender: Yup.string().required("This field is required"),
+  gender: Yup.string().required("Gender is required"),
   dateOfBirth: Yup.string()
-  .matches(/^(0?[1-9]|1[0-2])\/(0?[1-9]|1\d|2\d|3[01])\/(19|20)\d{2}$/, "Invalid Date, must be in MM/DD/YYYY")
-  .required("This field is required"),
+    .matches(
+      /^(0?[1-9]|1[0-2])\/(0?[1-9]|1\d|2\d|3[01])\/(19|20)\d{2}$/,
+      "Invalid Date, must be in MM/DD/YYYY"
+    )
+    .required("Date of birth is required"),
   email: Yup.string()
     .email("Email is invalid")
-    .required("This field is required"),
+    .required("Email is required"),
   phoneNumber: Yup.string()
     .matches(/((\+639)|09)\d{9}/, "Mobile phone number is not valid")
     .max(13, "Mobile phone number is too long")
-    .required("This field is required"),
+    .required("Mobile Phone is required"),
   completeAddress: Yup.string()
     .max(80, "Must be 50 characters or less")
     .min(20, "Must be 20 characters or more")